+++ mops/abstraction/driver_wrapper_abc.py
@@ -44,6 +44,8 @@ class DriverWrapperABC(ABC):
     is_simulator: bool = False
     is_real_device: bool = False
 
+    is_cdp: bool = False
+
     browser_name: str | None = None
 
     @cached_property