+++ mops/base/driver_wrapper.py
@@ -128,6 +128,8 @@ class DriverWrapper(InternalMixin, Logging, DriverWrapperABC):
     is_simulator: bool = False
     is_real_device: bool = False
 
+    is_cdp: bool = False
+
     browser_name: str | None = None
 
     def __new__(cls, *args: Any, **kwargs: Any) -> Self: